requirements.txt
semantic_version
sh
pyyaml